7.1.0:  rollout of 1.6.2 code

+++ src/me/sablednah/MobHealth/API/MobHealthAPI.java
@@ -250,7 +250,7 @@ public class MobHealthAPI {
             MonsterApocalypse ma = (MonsterApocalypse) plugin.getServer().getPluginManager().getPlugin("Monster Apocalypse");
             healthmanager MAHealthManager = ma.getHealthManager();
             if (MAHealthManager != null) {
-                targetHealth = MAHealthManager.getmobhp(targetMob);
+                targetHealth = (int) MAHealthManager.getmobhp(targetMob);
             }
             MAHealthManager = null;
             ma = null;
